requirements.txt
marqo==0.5.4
streamlit==1.13.0
pandas==1.5.0
python-dotenv==0.21.0
beautifulsoup4==4.11.1
selenium==4.6.0
